+++ CHANGES.rst
@@ -48,6 +48,7 @@ Improvements
   add an extra column with this information to Excel/CSV exports (:pr:`5330`)
 - Add option to delete persons from the event if they have no roles or other ties
   to the event anymore (:issue:`5294`, :pr:`5313`)
+- Allow events to be favorited (:issue:`1662`, :pr:`5338`, thanks :user:`Leats`)
 
 Bugfixes
 ^^^^^^^^